# docs/swagger.yaml
info:
  contact: {}
paths:
  /login:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 公共方法
  /problem-detail:
    get:
      parameters:
      - description: identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 问题详情
      tags:
      - 公共方法
  /problem-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: problem_identity
        in: query
        name: problemIdentity
        type: string
      - description: user_identity
        in: query
        name: userIdentity
        type: string
      - description: status
        in: query
        name: status
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 提交列表
      tags:
      - 公共方法
  /register:
    post:
      parameters:
      - description: username
        in: formData
        name: username
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      - description: mail
        in: formData
        name: mail
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - 公共方法
  /send-email:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 发送验证码
      tags:
      - 公共方法
  /user-detail:
    get:
      parameters:
      - description: identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 用户信息
      tags:
      - 公共方法
  /user-rank-list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: '{"code":200,"data":""}'
          schema:
            type: string
      summary: 用户排行榜
      tags:
      - 公共方法
swagger: "2.0"